refactor(): console overlay logic into a hook
